requirements.txt
pytest_allure_adaptor==1.7.10
requests==2.21.0
seleniumwrapper==0.5.4
pytest==3.1.0